+++ NEWS.md
@@ -1,5 +1,14 @@
 # dialr (development version)
 
+## New functions
+
+* Retrieve carrier name with `get_carrier()` (#6).
+
+* Add geocoding functions `get_geocode()` and `get_timezone()` (#5).
+
+## Minor changes
+
+* Rename "One shot methods" vignette to remove redundant "dialr".
 
 # dialr 0.2.1